@@ -1,6 +1,6 @@
 {
   "name": "rebass",
-  "version": "0.2.4",
+  "version": "0.2.5",
   "description": "Configurable React Stateless Functional UI Components",
   "author": "Brent Jackson",
   "license": "MIT",
@@ -12,7 +12,7 @@
     "test": "standard && mocha test --compilers js:babel-register"
   },
   "peerDependencies": {
-    "react": "^0.14.0"
+    "react": "^0.14.0 || ^15.0.0"
   },
   "devDependencies": {
     "babel-cli": "^6.5.1",
@@ -24,8 +24,8 @@
     "expect": "^1.14.0",
     "mocha": "^2.4.5",
     "onchange": "^2.0.0",
-    "react": "^0.14.7",
-    "react-addons-test-utils": "^0.14.7",
+    "react": "^15.0.1",
+    "react-addons-test-utils": "^15.0.1",
     "react-component-permutations": "^1.0.0-beta2",
     "react-docgen": "^2.7.0",
     "react-element-to-jsx-string": "^2.4.0",
