@@ -8,6 +8,7 @@ import Base from './Base'
 
 const LinkBlock = ({
   Component,
+  is,
   ...props
 }, { rebass }) => {
   const sx = {
@@ -15,6 +16,7 @@ const LinkBlock = ({
     textDecoration: 'none',
     color: 'inherit'
   }
+  Component = Component || is
 
   return (
     <Base
@@ -27,11 +29,21 @@ const LinkBlock = ({
 
 LinkBlock.propTypes = {
   /** Root component - useful for use with react-router's Link component */
-  Component: React.PropTypes.node
+  is: React.PropTypes.oneOfType([
+    React.PropTypes.string,
+    React.PropTypes.object,
+    React.PropTypes.func
+  ]),
+  /** Alias for `is` prop */
+  Component: React.PropTypes.oneOfType([
+    React.PropTypes.string,
+    React.PropTypes.object,
+    React.PropTypes.func
+  ])
 }
 
 LinkBlock.defaultProps = {
-  Component: 'a'
+  is: 'a'
 }
 
 LinkBlock.contextTypes = {
